server/requirements.txt
flask==3.1.2
flask-jwt-extended==4.7.1
pymongo==4.14.1
bcrypt==4.3.0
python-dotenv==1.1.1
reportlab==4.4.3
requests==2.32.5
gunicorn==22.0.0